+++ app/components/landing/LandingCTA.tsx
@@ -1,5 +1,5 @@
 // TODO: Phase out and delete once /home and /protocol/[id] adopt ConkaCTAButton.
-import { FUNNEL_URL } from "@/app/lib/landingConstants";
+import { BYO_URL } from "@/app/lib/landingConstants";
 import Button from "@/app/components/Button";
 
 /**
@@ -8,7 +8,7 @@ import Button from "@/app/components/Button";
  */
 export default function LandingCTA({
   children,
-  href = FUNNEL_URL,
+  href = BYO_URL,
   className = "",
 }: {
   children: React.ReactNode;